refactor(breadcrumb-link): rename index, allocate story
